requirements.txt
selenium
webdriver-manager
pyinstaller
pywin32
pycryptodomex
